Code modularized
